python_api_pipeline/requirements.txt

requests
python-dotenv
sqlalchemy>=2.0
openai
